Files for registration
The register.html file is a basic form to gather information, which is posted to the connect.php file. That file then connects to the database, inserts the records into the database, including encrypting the password. Currently, for debugging, it then shows some debugging output and the contents of the database via a select.