Added cookies to these and xfer
All of these uploaded files, excluding getproperties.php will do a header type php change of page to getproperties.php, with cookies set to show who the user is. This session will live for 4 hours.  The login php (conn2.php) will transfer back to login.html if the login username/password fails. It now shows an alert when the password and username do not match.