Merged files with Briannas
The files were merged with Brianna's. The login.html was modified to the point where it doesn't look as good. It will need Brianna to go over it again. The register.html (previously known as signup.html) had to be completely replaced. The javascript code was preventing the password and confirm password being confirmed as the same.

The mechanism to call connect.php and conn2.php had to be changed - they can only accept POSTs.The files were merged with Brianna's.